Added exceptions to service classes

+++ src/main/java/com/example/top/service/DimensionsService.java
@@ -1,6 +1,7 @@
 package com.example.top.service;
 
 import com.example.top.entity.order.Dimensions;
+import com.example.top.exception.UnknownIdException;
 import com.example.top.repository.DimensionsRepository;
 import lombok.extern.java.Log;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,14 +17,11 @@ public class DimensionsService {
     private DimensionsRepository repository;
 
     public void saveDimensions(Dimensions dimensions) {
-        if (dimensions == null) {
-            log.severe("Cannot add null as dimensions");
-            return;
-        }
+        if (dimensions == null) throw new IllegalArgumentException("The object 'Dimensions' cannot be null");
 
         repository.save(dimensions);
 
-        log.info("Dimensions with the name '" + dimensions.getName() + "' is saved");
+        log.info("Dimensions '" + dimensions.getName() + "' is saved");
     }
 
     public List<Dimensions> findAllDimensions() {
@@ -34,27 +32,25 @@ public class DimensionsService {
     }
 
     public Dimensions getDimensions(Long id) {
-        var optDimensions = repository.findById(id);
+        if (id == null) throw new IllegalArgumentException("Id cannot be null");
 
-        if (optDimensions.isEmpty()) {
-            log.severe("No dimensions found with the id '" + id + "'");
-            return null;
-        }
+        var optDimensions = repository.findById(id);
+        if (optDimensions.isEmpty()) throw new UnknownIdException("No dimensions found with the id '" + id + "'");
 
         var dimensions = optDimensions.get();
-        log.info("Dimensions with the name '" + dimensions.getName() + "' has been retrieved");
+        log.info("Dimensions '" + dimensions.getName() + "' has been retrieved");
         return dimensions;
     }
 
     public void deleteDimensions(Long id) {
+        if (id == null) throw new IllegalArgumentException("Id cannot be null");
+
         var dimensions = getDimensions(id);
-        if (dimensions == null) {
-            log.severe("Cannot delete the dimensions with the id '" + id + "' which doesn't exists");
-            return;
-        }
+        if (dimensions == null)
+            throw new IllegalStateException("Cannot delete the dimensions with the id '" + id + "' which doesn't exists");
 
         repository.deleteById(id);
 
-        log.info("Dimensions with the name '" + dimensions.getName() + "' has been deleted");
+        log.info("Dimensions '" + dimensions.getName() + "' has been deleted");
     }
 }
